requirements.txt
flask    
flask-CORS    
numpy    
cython    
PyMCubes   
scikit-image    
https://storage.googleapis.com/tensorflow/mac/tensorflow-0.9.0rc0-py3-none-any.whl
sympy

